==> 23_report_weight_analysis_mitt.qmd <==
---
title: "SPAN 2 Clot Study - Weight"
subtitle: "mITT"
author: "Marcio Diniz, Shu-Hsiang Wang, Karni Bedirian"
date: "`r format(Sys.time(), '%d %B, %Y')`"
format: 
  html:
    toc: true
    toc-depth: 5
    toc-location: left
    column: page-right
    self-contained: true
    echo: true
    code-fold: true
    warning: false
    message: false
---

```{r}
source("00_utils.R") 
load("results/span2_clotstudy_p3_weight_analysis_mitt.Rdata")
```

# Tables

::: {.panel-tabset}
## Overall


```{r}
tab <- tables$outcome$overall

nt_format(tab) |>
  tab_header(
    title = "Descriptive Analysis - Weight"
  ) 
```


## Site
```{r}
tab <- tables$outcome$site

nt_format(tab) |>
  tab_header(
    title = "Descriptive Analysis - Weight"
  ) 

```

## Sex
```{r}
tab <- tables$outcome$sex

nt_format(tab) |>
  tab_header(
    title = "Descriptive Analysis - Weight"
  ) 
```

## Treatment
```{r}
tab <- tables$outcome$txas_reperfusion

nt_format(tab) |>
  tab_header(
    title = "Descriptive Analysis - Weight"
  ) 
```

## Clot Length
```{r}
tab <- tables$outcome$clot_length

nt_format(tab) |>
  tab_header(
    title = "Descriptive Analysis - Weight"
  ) 
```

:::

# Plots 

::: {.panel-tabset}

## Overall

::: {.panel-tabset}

### Treatment

```{r}
#| fig.height: 7
#| fig-cap: "Mean \u00b1  SE"
plots$overall$txas_reperfusion
```

### Clot Length

```{r}
#| fig.height: 7
#| fig-cap: "Mean \u00b1  SE"
plots$overall$clot_length
```

:::

## CL3

```{r}
#| fig.height: 7
#| fig-cap: "Mean \u00b1  SE"
plots$cl3$txas_reperfusion
```

## CL4

```{r}
#| fig.height: 7
#| fig-cap: "Mean \u00b1  SE"
plots$cl4$txas_reperfusion
```

:::

# Test of Hypotheses

::: {.panel-tabset}

## Overall

::: {.panel-tabset}

### Interaction

```{r}
gt(tables$lmer$overall$interaction) |>
  tab_header(
    title = "Interaction"
  ) |>
  tab_footnote(
    footnote = footnote$pvals,
    locations = cells_column_labels(columns = "p value")
  ) 
```

### Treatment

```{r}
gt(tables$lmer$overall$txas_reperfusion) |>
  tab_header(
    title = "Comparison at each time point"
  ) |>
  tab_footnote(
    footnote = footnote$ci,
    locations = 
      cells_column_labels(columns = "Difference (95%CI)")
  ) |>
  tab_footnote(
    footnote = footnote$pvals,
    locations = cells_column_labels(columns = "p value")
  ) 
```

### Clot Length

```{r}
gt(tables$lmer$overall$clot_length) |>
  tab_header(
    title = "Comparison at each time point"
  ) |>
  tab_footnote(
    footnote = footnote$ci,
    locations = 
      cells_column_labels(columns = "Difference (95%CI)")
  ) |>
  tab_footnote(
    footnote = footnote$pvals,
    locations = cells_column_labels(columns = "p value")
  ) 
```

:::

## CL3

::: {.panel-tabset}

### Interaction 

```{r}
gt(tables$lmer$cl3$interaction) |>
  tab_header(
    title = "Interaction"
  ) |>
  tab_footnote(
    footnote = footnote$pvals,
    locations = cells_column_labels(columns = "p value")
  ) 
```

### Treatment

```{r}
gt(tables$lmer$cl3$txas_reperfusion) |>
  tab_header(
    title = "Comparison at each time point"
  ) |>
  tab_footnote(
    footnote = footnote$ci,
    locations = 
      cells_column_labels(columns = "Difference (95%CI)")
  ) |>
  tab_footnote(
    footnote = footnote$pvals,
    locations = cells_column_labels(columns = "p value")
  ) 
```

:::

## CL4

::: {.panel-tabset}

### Interaction 

```{r}
gt(tables$lmer$cl4$interaction) |>
  tab_header(
    title = "Interaction"
  ) |>
  tab_footnote(
    footnote = footnote$pvals,
    locations = cells_column_labels(columns = "p value")
  ) 
```

### Treatment

```{r}
gt(tables$lmer$cl4$txas_reperfusion) |>
  tab_header(
    title = "Comparison at each time point"
  ) |>
  tab_footnote(
    footnote = footnote$ci,
    locations = 
      cells_column_labels(columns = "Difference (95%CI)")
  ) |>
  tab_footnote(
    footnote = footnote$pvals,
    locations = cells_column_labels(columns = "p value")
  ) 
```

:::
